# Conflicts:
#	src/components/Stingray.tsx
#	src/pages/stingray.tsx

+++ db/init/_init_tables.sql
@@ -78,15 +78,16 @@ SET time_zone = '+00:00';
 CREATE TABLE stingray
 (
     `stingray_id` int         AUTO_INCREMENT,
-    `longitude`   int         NOT NULL,
-    `latitude`    int         NOT NULL,
+    `longitude`   float       NOT NULL,
+    `latitude`    float       NOT NULL,
     `pitch`       float       NOT NULL,
     `roll`        float       NOT NULL,
     `time`        timestamp   NOT NULL,
     `tempratur`   int         NOT NULL,   
-    primary key (`stingray_id`),
-   
+    primary key (`stingray_id`), 
+
 ) ENGINE = InnoDB
   DEFAULT CHARSET = utf8mb4
   COLLATE = utf8mb4_0900_ai_ci;
+
